backend/requirements.txt
# SafeDriveApp Backend Requirements

# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
starlette>=0.27.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Authentication and security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0

# Environment and configuration
python-dotenv>=1.0.0

# Image processing and computer vision
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0

# File handling and utilities
python-multipart>=0.0.6
aiofiles>=23.0.0

# HTTP client (if needed)
httpx>=0.25.0
requests>=2.31.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0